#82787 Get subscriber details via API
